+++ src/together/lib/cli/api/beta/clusters/list_regions.py
@@ -2,9 +2,11 @@ import json as json_lib
 from typing import Any, Dict, List
 
 import click
+from rich import print
 from tabulate import tabulate
 
 from together import Together
+from together.lib.cli._track_cli import auto_track_command
 from together.lib.cli.api._utils import handle_api_errors
 
 
@@ -16,6 +18,7 @@ from together.lib.cli.api._utils import handle_api_errors
 )
 @click.pass_context
 @handle_api_errors("Clusters")
+@auto_track_command
 def list_regions(ctx: click.Context, json: bool) -> None:
     """List regions"""
     client: Together = ctx.obj
@@ -27,13 +30,19 @@ def list_regions(ctx: click.Context, json: bool) -> None:
     else:
         data: List[Dict[str, Any]] = []
         for region in response.regions:
+            driver_versions: list[str] = []
+            for driver_version in region.driver_versions:
+                driver_versions.append(
+                    f"[dim]NVIDIA Driver:[/dim] [blue]{driver_version.nvidia_driver_version}[/blue] [dim]CUDA Version:[/dim] [blue]{driver_version.cuda_version}[/blue]"
+                )
+
             data.append(
                 {
                     "Name": region.name,
                     "Supported GPU Types": ", ".join(region.supported_instance_types)
                     if region.supported_instance_types
                     else "",
-                    "Driver Versions": ", ".join(region.driver_versions) if region.driver_versions else "",
+                    "Driver Versions": "\n".join(driver_versions) if driver_versions else "",
                 }
             )
-        click.echo(tabulate(data, headers="keys", tablefmt="grid"))
+        print(tabulate(data, headers="keys", tablefmt="grid"))